@@ -1,9 +1,10 @@
 package com.example.mulyani.resepmakanan;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class RincianActivity extends AppCompatActivity {
 
